+++ src/particles/video/component.tsx
@@ -71,6 +71,7 @@ export interface IVideoProps extends IComponentProps<IVideoTheme> {
   shouldMute?: boolean;
   shouldLoop?: boolean;
   isLazyLoadable?: boolean;
+  ipfsPrefix?: string;
   onEnded?: () => void;
   onPlayed?: () => void;
   onPaused?: () => void;
@@ -83,6 +84,7 @@ export const Video = (props: IVideoProps): React.ReactElement => {
   const shouldShowControls = props.shouldShowControls != null ? props.shouldShowControls : true;
   const width = props.width ? props.width : props.isFullWidth ? '100%' : 'auto';
   const height = props.height ? props.height : props.isFullHeight ? '100%' : 'auto';
+  const source = props.source.startsWith('ipfs://') ? props.source.replace('ipfs://', props.ipfsPrefix ?? 'https://ipfs.io/ipfs/') : props.source;
 
   const onEnded = (): void => {
     if (props.onEnded) {
@@ -123,7 +125,8 @@ export const Video = (props: IVideoProps): React.ReactElement => {
       onEnded={onEnded}
       onPlay={onPlayed}
       onPause={onPaused}
-      src={props.source}
+      src={source}
+      crossOrigin='anonymous'
     >
       {props.alternativeText}
     </StyledVideo>